requirements.txt
pipeit
uvicorn[standard]
numpy 
pillow
aiohttp
fastapi
ThreadPoolExecutorPlus
aiofiles
janome
mecab-python3
SudachiPy
gTTS
pyttsx3
manga-ocr
python-multipart
beautifoulsoup4
selenium
wget